+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Allow set comparison between groups and single platform.
+
 ## [1.3.1 (2024-09-18)](https://github.com/kdeldycke/extra-platforms/compare/v1.3.0...v1.3.1)
 
 - Fix conflicting detection heuristics for Linux distributions. Closes #72.